poprawne suswanie komentzry

+++ Controller/PostController.php
@@ -232,22 +232,43 @@ class PostController extends MainController
         }
 
 
-        if ((!empty($_POST['delete'])) && ($this->oModel->delete($post_id))) {
+        /* if ((!empty($_POST['delete'])) && ($this->oModel->delete($post_id))) {
 
-            header('Location: ' . MAIN_PAGE);
-            if (!empty($_SESSION['PostErrorMsg'])) {
-                unset($_SESSION['PostErrorMsg']);
+             header('Location: ' . MAIN_PAGE);
+             if (!empty($_SESSION['PostErrorMsg'])) {
+                 unset($_SESSION['PostErrorMsg']);
+             }
+             $_SESSION['PostSuccessMsg'] = 'Hurray!! The post has been deleted.';
+         } else {
+             header('Location: ' . MAIN_PAGE);
+             if (!empty($_SESSION['PostSuccessMsg'])) {
+                 unset($_SESSION['PostSuccessMsg']);
+             }
+             $_SESSION['PostErrorMsg'] = 'The post cant be deleted.';
+         }*/
+
+        //when comments exists
+        if ($this->CommentModel->getByIdCheck($post_id)) {
+
+            if (!empty($_POST['delete'])
+                && $this->oModel->delete($post_id)
+                && $this->CommentModel->delete($post_id)) {
+
+                header('Location: ' . MAIN_PAGE);
+            } else {
+                exit('Whoops! Post cannot be deleted.');
             }
-            $_SESSION['PostSuccessMsg'] = 'Hurray!! The post has been deleted.';
-        } else {
-            header('Location: ' . MAIN_PAGE);
-            if (!empty($_SESSION['PostSuccessMsg'])) {
-                unset($_SESSION['PostSuccessMsg']);
-            }
-            $_SESSION['PostErrorMsg'] = 'The post cant be deleted.';
+
         }
+        //when comments not exists
+        if (!$this->CommentModel->getByIdCheck($post_id)) {
+            if ((!empty($_POST['delete'])) && ($this->oModel->delete($post_id))) {
+                header('Location: ' . MAIN_PAGE);
+            } else {
+                exit('Whoops! Post cannot be deleted.');
+            }
 
-        //  exit('Whoops! Post cannot be deleted.');
+        }
     }
